Example of Q-Learning RF Agent
The python script will train an agent using Q-Learning tables to explore and memorize a way out of a maze. It does not know the maze in advanced but learns details of the maze through exploration. The user can change the maze and various starting parameters including exploration rate, discount value, learning rate, maximum number of steps before giving up and maximum number of episodes before stopping the training session.  One can also load a previously saved q-table from successful runs but it will only work if the new maze is of the same dimensions, there it not currently a convertor, simply put them in the starting_q_tables folder and set starting parameters to point there. Otherwise comment out in the second cell to do a "fresh run".  Maze 6 and 7 are of the same dimensions and the provided "starting q-table" has been trained on both of them. You can look at the run history to see whichever run was most successful by looking at the number of steps taken, ignore negative values as those are fails.  Videos of successful runs can be found in the frames folder. If you terminate a run early you can manually convert these frames to videos with the final cell of the script.   You can change frequency of image/video saves in the starting parameters section at the start of the script.  The trails run significantly faster with fewer saves.